IMPORTANT: for the next piece to work we definitely also need to expose RabbitMQ's broker port (5672), so stop and restart RabbitMQ in docker as follows:
docker rm -f springcloud-rabbit
docker run -d -p 15672:15672 -p 5672:5672 --hostname rabbithost --name springcloud-rabbit rabbitmq:3-management

With both the Intake (port 8082) and Sink (port 8080) processes running and hosted in their own Tomcat instances, send a POST request with a 3-element CSV raw string (e.g. Station1,PLATE01,20160101_1200) in the body to:

	http://localhost:8082/tasks.

You will observe data in some new tables in the tasklogs database.

Here's how that happened:

The TaskController defines an HTTP endpoint for receiving the above request.

This uses an auto-wired TaskProcessor to emit a publication.

The TaskProcessor component is bound and wired to a Source which is configured to a Rabbit (port 5672) message queue named tasktopic.

A task request (not YET a task) is sent to the tasktopic queue.  The message contains the fully-qualified URL of the Maven-Installed Task JAR plus any parameters supplied in the POST.

The Sink application has virtually zero code but it's configured to subscribe to the tasktopic queue and it instinctively knows what to do with each message - it spawns a Task process, passing in the supplied args.

The Task JAR defines a main() entry point so it represents a standalone process.  It's configured to dump it's state (e.g. process args) to MySQL upon each invocation.

NOTE: Was struggling to get this to work reliably using Spring Cloud version Camden.SR3, so reverted to Brixton.SR6.